refactor: rename parameter_docs to docs
The module documents how to address the API, not parameters specifically, so the
narrower name would have to change the moment a tool-level note is added.

Its docstring now distinguishes it from docs_search.py, which searches the
Appwrite product documentation for the user rather than describing the API.
